revert: restore componentService in elementService
