+++ libs/frontend/domain/app/src/store/app.service.ts
@@ -2,12 +2,10 @@ import {
   IApp,
   IAppDTO,
   IAppService,
-  IComponentService,
   ICreateAppDTO,
   IElementService,
   IPageBuilderAppProps,
   IPageService,
-  isComponentDTO,
   IStoreService,
   IUpdateAppDTO,
 } from '@codelab/frontend/abstract/core'
@@ -45,7 +43,6 @@ export class AppService
     deleteModal: prop(() => new AppModalService({})),
 
     _elementService: prop<Ref<IElementService>>(),
-    _componentService: prop<Ref<IComponentService>>(),
     pageService: prop<IPageService>(),
     storeService: prop<IStoreService>(),
   })
@@ -54,11 +51,6 @@ export class AppService
   @computed
   get elementService() {
     return this._elementService.current
-  }
-
-  @computed
-  get componentService() {
-    return this._componentService.current
   }
 
   /**
@@ -89,14 +81,6 @@ export class AppService
     }
 
     const elements = [page.rootElement, ...page.rootElement.descendantElements]
-
-    const components = elements
-      .map((v) => v.parentComponent || v.renderComponentType)
-      .filter(isComponentDTO)
-
-    components.forEach((component) =>
-      this.componentService.writeCache(component),
-    )
 
     const pageElements = elements.map((element) =>
       this.elementService.writeCache(element),
