+++ libs/frontend/domain/renderer/src/test/setup/setupTest.ts
@@ -13,6 +13,7 @@ import {
   CUSTOM_TEXT_PROP_KEY,
   Element,
   ElementService,
+  elementServiceContext,
   ElementTree,
 } from '@codelab/frontend/domain/element'
 import { Prop } from '@codelab/frontend/domain/prop'
@@ -25,10 +26,6 @@ import {
   typeRef,
   TypeService,
 } from '@codelab/frontend/domain/type'
-import {
-  componentRef,
-  elementServiceContext,
-} from '@codelab/frontend/presenter/container'
 import { PrimitiveTypeKind } from '@codelab/shared/abstract/codegen'
 import { IAtomType } from '@codelab/shared/abstract/core'
 import { frozen, objectMap, unregisterRootStore } from 'mobx-keystone'
@@ -127,7 +124,7 @@ export const setupTestForRenderer = (pipes: Array<RenderPipeClass> = []) => {
       customCss: '',
       guiCss: '',
       atom: atomRef(data.textAtom.id),
-      parentComponent: componentRef(data.componentToRender),
+      parentComponent: data.componentToRender,
       props: new Prop({
         id: v4(),
         data: frozen({
@@ -140,7 +137,7 @@ export const setupTestForRenderer = (pipes: Array<RenderPipeClass> = []) => {
     data.componentInstanceElementToRender = new Element({
       id: v4(),
       name: '01',
-      renderComponentType: componentRef(data.componentToRender),
+      renderComponentType: data.componentToRender,
       props: new Prop({
         id: v4(),
         data: frozen({
@@ -181,6 +178,12 @@ export const setupTestForRenderer = (pipes: Array<RenderPipeClass> = []) => {
     }`,
     })
 
+    const componentService = new ComponentService({
+      components: objectMap([
+        [data.componentToRender.id, data.componentToRender],
+      ]),
+    })
+
     data.rootStore = new RenderTestRootStore({
       typeService: new TypeService({
         types: objectMap([
@@ -189,11 +192,7 @@ export const setupTestForRenderer = (pipes: Array<RenderPipeClass> = []) => {
           [data.reactNodeType.id, data.reactNodeType],
         ]),
       }),
-      componentService: new ComponentService({
-        components: objectMap([
-          [data.componentToRender.id, data.componentToRender],
-        ]),
-      }),
+      componentService: componentService,
       atomService: new AtomService({
         atoms: objectMap([
           [data.divAtom.id, data.divAtom],
