+++ libs/frontend/model/infra/mobx/src/create-root-store.ts
@@ -60,7 +60,6 @@ export const createRootStore = ({ user }: RootStoreData) => {
 
   const appService = new AppService({
     _elementService: elementServiceRef(elementService),
-    _componentService: componentServiceRef(componentService),
     pageService,
     storeService,
   })
